\name{ssfit-package}
\alias{ssfit-package}
\alias{ssfit}
\docType{package}
\title{
Fitting of parametric models using summary statistics
}
\description{
Fits complex parametric models without likelihoods, using the method proposed by Cox and Kartsonaki (2012).
}
\details{
\tabular{ll}{
Package: \tab ssfit\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2013-08-21\cr
Depends: survey
License: \tab GPL (>= 2)\cr
}

See \code{fit.model}.
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}
\author{
Christiana Kartsonaki

Maintainer: Christiana Kartsonaki <christiana.kartsonaki@gmail.com>
}
\references{
Cox, D. R. and Kartsonaki, C. (2012). The fitting of complex parametric models. \emph{Biometrika}, \bold{99} (3): 741--747.
}

%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
%\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
